A few days ago, I upgraded my devices to iOS 7.1 (11D5145e), and to Xcode Version 5.1 (5B103i) aka Xcode51-Beta5. Two things broke: [1] the “Storyboard/UITableView/UITableViewCell” connection; and [2] how to determine when a discovered document’s object graph becomes available.
[1] Xcode51-Beta5 discovered the Storyboard snafu; it’s just pickier.
I just ditched the DocumentCell class and reverted to plain UITableViewCell.

[2] Suddenly, the previous version posted to GitHub just wouldn’t permit the user to inspect a discovered document; its row/cell remained grey (userInteractionEnabled=NO).
I had relied on NSPersistentStoreDidImportUbiquitousContentChangesNotification to signal when a document’s object graph became available.
After several hours of peeking and poking, I discovered a useful signal is now:

	NSConcreteNotification 0x165ea720 {name =
	com.apple.coredata.ubiquity.importer.didfinishimport;
	object = <NSPersistentStoreCoordinator: 0x16695b40>;
	userInfo = {
             deleted = "{(\n)}";
             inserted = "{(\n    0x16586180 <x-coredata://E24729CE-4C32-4961-A228-005264DCA31B/ModelVersion/p2>,\n    0x165de350 <x-coredata://E24729CE-4C32-4961-A228-005264DCA31B/TextEntry/p2>\n)}";
             updated = "{(\n)}";
             }}

I’ve adopted a work-around.
See:
	-[DocumentsListController+Making receivedNotification:document:]
and
	-[NSDictionary+NPAssisting isDocumentViewable].

+++ MultiDocument/DetailViewController.m
@@ -186,7 +186,7 @@
         }
         default:
         {
-            NSAssert( NO, @"Bogus count of fetched objects = %d", fObjs.count);
+            NSAssert( NO, @"Bogus count of fetched objects = %lu", (unsigned long)fObjs.count);
             break;
         }
     }
